LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-en-US.html

<!DOCTYPE html><!-- webkit-test-runner [ language=en-US ] -->
<html>
<head>
<script src="../../../../../resources/js-test.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "en");
</script>
</body>
</html>
